requirements.txt
auth==0.5.3
Django==4.0.4
django_filter==21.1
djangorestframework==3.13.1
